# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go-Secure-Gateway API
  version: 1.0.0
  description: |
    API surface exposed by the gateway. Use this spec to build an automated
    test framework: import it into Postman/Bruno, generate a client, or drive
    schema-based assertions.

    The backends behind the protected routes are controllable test doubles:
      - any request may carry `?status=<code>` to force an HTTP status
        (e.g. `?status=503` to trip the circuit breaker), and/or
      - `?delay=<duration>` (e.g. `?delay=2s`) to inject latency.
servers:
  - url: http://localhost:8080
    description: Gateway (direct)
  - url: http://localhost:8088/api
    description: Via the demo frontend's nginx proxy
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /debug/token:
    get:
      summary: Mint a development JWT (only when debug=true)
      security: []
      responses:
        "200":
          description: A signed token
          content:
            application/json:
              schema:
                type: object
                properties: { token: { type: string } }
        "404": { description: Debug mode disabled }
...